+++ app/Item.php
@@ -32,8 +32,9 @@ class Item extends Model
             'pFsense' => \App\SupportedApps\Pfsense::class,
             'Pihole' => \App\SupportedApps\Pihole::class,
             'Plex' => \App\SupportedApps\Plex::class,
-            'UniFi' => \App\SupportedApps\Unifi::class,
             'Portainer' => \App\SupportedApps\Portainer::class,
+            'UniFi' => \App\SupportedApps\Unifi::class,
+            'Sabnzbd' => \App\SupportedApps\Sabnzbd::class,
         ];
     }
     public static function supportedOptions()
@@ -58,13 +59,39 @@ class Item extends Model
         $view = null;
         if(isset($this->description) && !empty($this->description)){
             $output = json_decode($this->description);
+            $output = is_object($output) ? $output : new \stdClass();
             if(isset($output->type) && !empty($output->type)) {
                 $class = $output->type;
                 $sap = new $class();
                 $view = $sap->configDetails();
+                $output->view = $view;
             }
-            $output->view = $view;
+            if(!isset($output->dataonly)) $output->dataonly = '0';
+            
         }
         return (object)$output;
+    }
+    public static function checkConfig($config)
+    {
+        if(empty($config)) {
+            $config = null;
+        } else {
+            $store = false;
+            //die(var_dump($config));
+            foreach($config as $key => $check) {
+                if($key == 'type') continue;
+                if($key == 'dataonly') continue;
+                if(!empty($check) && $check != '0') {
+                    $store = true;
+                    break;
+                }
+            }
+            //die(var_dump($store))
+            
+            $config['enabled'] = ($store) ? true : false;
+            $config = json_encode($config);
+        }
+        return $config;
+
     }
 }