+++ app/SupportedApps/Contracts/Livestats.php
@@ -1,0 +1,11 @@
+<?php namespace App\SupportedApps\Contracts;
+
+interface Livestats {
+
+    public function configDetails();
+
+    public function testConfig();
+
+    public function executeConfig();
+    
+}