+++ resources/lang/en/app.php
@@ -59,6 +59,8 @@ return [
     'apps.username' => 'Username',
     'apps.password' => 'Password',
     'apps.config' => 'Config',
+    'apps.apikey' => 'Api Key',
+    'apps.enable' => 'Enable',
 
     'url' => 'Url',
     'title' => 'Title',