+++ resources/lang/tr/auth.php
@@ -1,0 +1,19 @@
+<?php
+
+return [
+
+    /*
+    |--------------------------------------------------------------------------
+    | Authentication Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are used during authentication for various
+    | messages that we need to display to the user. You are free to modify
+    | these language lines according to your application's requirements.
+    |
+    */
+    
+    'failed' => 'Kimlik bilgileri doğru değil.',
+    'throttle' => 'Çok fazla girişim. :seconds saniye sonra tekrar deneyin.',
+    
+];