+++ resources/lang/tr/passwords.php
@@ -1,0 +1,22 @@
+<?php
+
+return [
+    
+    /*
+    |--------------------------------------------------------------------------
+    | Password Reset Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are the default lines which match reasons
+    | that are given by the password broker for a password update attempt
+    | has failed, such as for an invalid token or invalid new password.
+    |
+    */
+    
+    'password' => 'Şifre en az altı karakter olmalı ve onaylamasına uymalıdır.',
+    'reset' => 'Şifreniz sıfırlandı!',
+    'sent' => 'Şifre sıfırlama bağlantısı eposta adresinize yollandı!',
+    'token' => 'Şifre sıfırlama simgesi geçerli değil.',
+    'user' => "Adresle ilişkili kullanıcı adı bulunamadı.",
+    
+];