+++ resources/views/items/form.blade.php
@@ -6,32 +6,28 @@
                 <a href="{{ route('items.index') }}" class="button"><i class="fa fa-ban"></i><span>{{ __('app.buttons.cancel') }}</span></a>
             </div>
         </header>
-        <div class="create">
+        <div id="create" class="create">
             {!! csrf_field() !!}
-            <!--<div class="input">
-                <label>Application name</label>
-                {!! Form::select('supported', \App\Item::supportedOptions(), array('placeholder' => 'Title','class' => 'form-control')) !!}
-            </div>-->
 
             <div class="input">
                 <label>{{ __('app.apps.application_name') }} *</label>
                 {!! Form::text('title', null, array('placeholder' => __('app.apps.title'), 'id' => 'appname', 'class' => 'form-control')) !!}
                 <hr />
                 <label>{{ strtoupper(__('app.url')) }}</label>
-                {!! Form::text('url', null, array('placeholder' => __('app.url'),'class' => 'form-control')) !!}
+                {!! Form::text('url', null, array('placeholder' => __('app.url'), 'id' => 'appurl', 'class' => 'form-control')) !!}
             </div>
             <div class="input">
                 <label>{{ __('app.apps.colour') }} *</label>
                 {!! Form::text('colour', null, array('placeholder' => __('app.apps.hex'),'class' => 'form-control color-picker')) !!}
                 <hr />
                 <label>{{ __('app.apps.pinned') }}</label>
+                {!! Form::hidden('pinned', '0') !!}
                 <label class="switch">
                     <?php
                     $checked = false;
                     if(isset($item->pinned) && (bool)$item->pinned === true) $checked = true;
                     $set_checked = ($checked) ? ' checked="checked"' : '';
-                    ?>
-                    {!! Form::hidden('pinned', '0') !!}
+                    ?>                   
                     <input type="checkbox" name="pinned" value="1"<?php echo $set_checked;?> />
                     <span class="slider round"></span>
                 </label>