+++ CHANGELOG.md
@@ -1,10 +1,41 @@
 # Release Notes
 
-## v1.1.2 (2018-02-05)
+## v1.3.0 (2018-02-05)
+
+### Added
+- guzzlehttp/guzzle as a dependency
+- Variable polling, so interval is increased when an app is idle and decreased when it's active
+- Turkish language translation
+- Added Sabnzbd enhanced application
+
+### Changed
+- Updated composer dependencies
+- Added live stats to Nzbget supported application  
+- Changed Pihole to an enhanced application
+- Changed NZBGet to an enhanced application
+
+### Fixed
+- Fixed autocomplete being hard to see
+
+
+## v1.2.0 (2018-02-07)
 
 ### Added
 - Translation support
 - Initial "Supported" application support
+- Finnish translation
+- Swedish translation
+- German translation
+- French translation
+- Spanish translation
+- Duplicati supported application
+- Emby supported application
+- Nzbget supported application
+- Pfsense supported application
+- Pihole supported application
+- Plex supported application
+- Portainer supported application
+- Unifi supported application
 
 ### Changed
 - button layout and behaviour